+++ agrest-cayenne/src/test/java/io/agrest/cayenne/GET_EntityOverlay_PerStackIT.java
@@ -11,7 +11,7 @@ import io.agrest.cayenne.unit.DbTest;
 import io.agrest.jaxrs.AgJaxrs;
 import io.agrest.meta.AgEntity;
 import io.agrest.meta.AgEntityOverlay;
-import io.agrest.runtime.AgBuilder;
+import io.agrest.runtime.AgRuntimeBuilder;
 import io.bootique.junit5.BQTestTool;
 import org.apache.cayenne.Cayenne;
 import org.apache.cayenne.query.SelectById;
@@ -32,7 +32,7 @@ public class GET_EntityOverlay_PerStackIT extends DbTest {
             .agCustomizer(GET_EntityOverlay_PerStackIT::addOverlay)
             .build();
 
-    private static AgBuilder addOverlay(AgBuilder builder) {
+    private static AgRuntimeBuilder addOverlay(AgRuntimeBuilder builder) {
 
         // creating an adhoc relationship between two persistent objects with a custom resolver
         AgEntityOverlay<E22> e22Overlay = AgEntity